requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
djangorestframework==3.14.0
idna==3.4
Pillow==9.5.0
pytz==2023.3
requests==2.28.2
urllib3==1.26.15
